From: hamogu <hgunther@mit.edu> Date: Thu, 7 Apr 2016 13:35:37 -0400 Subject: [PATCH] Replace xxxConstantARF functions with xxxARFfromarrays
Currently, the binning of the ARF determines the way that photons are
drawn from an energy spectrum. The ARF in fact sets the binning of the
spectrum, and photon energies are randomized within each bin - even if the
energy spectrum has a much finer grid than the ARF.
Thus, the xxxConstantARF methods that set exactly one ARF bin are useless
(unless you want a flat spectrum).
On the other hand, we also don't want to determine the binning of the ARF
hardcoded in SIMPUT because 0.1 keV might be sufficient for e.g. ROSAT
simulations, while a microcalorimeter or grating requires a much finer binning.

Thus, the new implementation allows the caller to input the ARF as arrays.
In contrast to the traditional methods that load the ARF from a file, the calling
code can make up any shape of ARF (including a constant) on the fly with any bin
width without writing an intermediate file.

 Changes to be committed:
	modified:   libsimput/arf.c
	modified:   libsimput/arf.h
	modified:   libsimput/datahandling.c
	modified:   libsimput/simput.h